fix: infrastructure alignment for Strat naming changes
- lib/backtest.py: Trade.strat_bonus and total_score → float (was int)
- lib/config.py: get_position_size/get_signal_strength_label type hints → float
- gcp/schema.sql: add ALTER TABLE migrations for production:
  - Rename premarket_analysis.strat_daily → strat_candle
  - Add prev_quarter_* columns to market_data_daily
  - Widen strat_combo VARCHAR(30) → VARCHAR(40) for new longer labels
- tests/test_config.py: update FTFC weight/timeframe assertions
- tests/test_data_loader.py: 'D' → '1d' in aggregate test

+++ lib/backtest.py
@@ -34,8 +34,8 @@ class Trade:
     entry_price: float
     direction: str  # 'CALL' or 'PUT'
     base_score: int
-    strat_bonus: int
-    total_score: int
+    strat_bonus: float
+    total_score: float
     position_size: float
     conditions_met: List[str]
     indicators_at_entry: Dict[str, float] = field(default_factory=dict)
@@ -633,7 +633,7 @@ class BacktestEngine:
         # --- Strat: FTFC score + ORB trend + filtering + bonus ---
         self._filter_counts['signals_evaluated'] += 1
 
-        strat_bonus = 0
+        strat_bonus = 0.0
         bar_ftfc_score = 0.0
         bar_orb_trend = 0
